requirements.txt
psycopg2-binary>=2.9
Flask>=2.3.5
requests